Use size_t for rd_file_fwrite_fortio

+++ lib/include/resdata/rd_file.hpp
@@ -32,7 +32,7 @@ int rd_file_get_size(const rd_file_type *rd_file);
 const char *rd_file_get_src_file(const rd_file_type *rd_file);
 rd_version_enum rd_file_get_rd_version(const rd_file_type *file);
 void rd_file_fwrite_fortio(const rd_file_type *ec_file, ERT::FortIO &fortio,
-                           int offset);
+                           size_t offset);
 
 int rd_file_get_phases(const rd_file_type *init_file);
